requirements.txt
# MCP SDK and core dependencies
mcp==1.1.2
pydantic==2.10.4
pydantic-settings==2.7.0

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Django integration
django==4.2.7
python-decouple==3.8

# AWS
boto3==1.35.82
botocore==1.35.82

# Monitoring and metrics
prometheus-client==0.21.1
psutil==6.1.1

# Async support
asyncio==3.4.3
aiohttp==3.11.11
websockets==14.1

# Utilities
python-dotenv==1.0.1
pytz==2024.2
python-dateutil==2.9.0.post0

# Alerting
requests==2.32.3
slack-sdk==3.34.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-mock==3.14.0

# Development
black==24.10.0
pylint==3.3.3
mypy==1.14.1